requirements.txt
# Python 3.10

django
djangorestframework
python-telegram-bot
django-dotenv